+++ src/features/pastelAddressBlock/PastelAddressBlock.module.css
@@ -1,0 +1,119 @@
+.well {
+  background-color: #000;
+  padding: 16px;
+}
+
+.accordionHeader:hover {
+  cursor: pointer;
+}
+
+.accordionHeader:focus {
+  outline: none;
+}
+
+.flexspacebetween {
+  display: flex;
+  justify-content: space-between;
+}
+
+.verticalflex {
+  flex-direction: column;
+}
+
+.marginleft {
+  margin-left: 16px;
+}
+
+.margintopSmall {
+  margin-top: 16px;
+}
+
+.margintoplarge {
+  margin-top: 24px;
+}
+
+.sublight {
+  color: #6a6a6a;
+}
+
+.padtopsmall {
+  padding-top: 4px;
+}
+
+.fixedfont {
+  font-family: monospace;
+}
+
+.breakword {
+  word-break: break-all;
+}
+
+.highlight {
+  color: #c3921f;
+}
+
+.yellow {
+  color: yellow;
+}
+
+.padrightsmall {
+  padding-right: 4px;
+}
+
+.small {
+  font-size: 12px;
+}
+
+.primarybutton {
+  background-color: #c3921f;
+  color: #000;
+  text-align: center;
+  font-weight: bold;
+  font-size: 14px;
+  padding-top: 8px;
+  padding-bottom: 8px;
+  padding-left: 16px;
+  padding-right: 16px;
+  border-radius: 4px;
+  border-color: #c3921f;
+  margin-left: 8px;
+  margin-right: 8px;
+}
+
+.primarybutton:disabled,
+.primarybutton[disabled] {
+  border: 1px solid #999;
+  background-color: #ccc;
+  color: #666;
+}
+
+.primarybutton:hover {
+  cursor: pointer;
+}
+
+.receiveQrcode {
+  padding: 8px;
+  background-color: white;
+}
+
+.buttomMarginTop {
+  margin-top: 10px;
+}
+
+.receiveblock {
+  margin-bottom: 16px;
+  display: flex;
+  flex-direction: column;
+}
+
+.lastAccessed {
+  display: inline-block;
+  margin-left: 10px;
+  font-size: 12px;
+  font-style: italic;
+  color: #fff;
+}
+
+.itemsCenter {
+  align-items: center;
+}
