+++ src/features/pastelAddressBlock/index.ts
@@ -1,0 +1,3 @@
+import PastelAddressBlock from './PastelAddressBlock'
+
+export default PastelAddressBlock
